Add search functionality from notification
Adds the ability to search for a product in the inventory directly from a notification. If the notification is related to food expiry or inventory, the app will open the `InventoryActivity` and automatically search for the related product. The product name is extracted from the notification message. If data is not null, it will be used as a search parameter.